Reworking civ3parse.py. Changed section name recognition logic
Also added hexdump function but not using it yet. The current
goal is to go back to basics and read all the headers/objects in.
But now that I have the hexdump function and horspool search I
should be able to better parse all sections from arbitrary game
saves. Well, not yet, but getting there.