About to try to add optional parameter to genericSection
So I can pass it the name if I've already consumed it with a horspool
search. Updates in this commit are parsing changes. I'm giving up
on the rest of the BIC and about to seek to the first LEAD. I'd
like to use genericSection on it and CNSL and any other section I
horspool skip to.