The parameter add was successful, but having trouble decoding...
Some data structures have a 4-byte ASCII value followed by a length,
but not all of them. I'm going to have to change strategies from
tring to figure the file out front-to-back and start comparing
features between different save files. Perhaps some data structures
are a static size and/or an array of a fixed size.