Added iterate and mystring sh scripts
iterate takes the files in a folder and determines if they are
compressed or uncompressed Civ3 saves, then if necessary pipes
the file through a decompressor (blast from zlib contrib, not
included here) to a specified file. mystrings is the file I've
specified, its aim is to count occurrences of headers. It's not
perfect but it helps.

LEAD appears 32 times per save, and ESPN 64 times. But some BICs
have occurrences of those strings, so I may have to start separating
the BIC from the rest of the game when counting and searching for
game data.