Have all my test saves parsing up to 2nd GAME
I now think this is the demarc between the embedded BIC and the
specific-game data.

TODO: The FLAV section in the Bic is different, and currently it
returns dummy data when dumpSelf is invoked although the data is
actually in the class instance

Also made iterate background the processes so everything goes in
parallel.